Fixed linting issues
