+++ pkg/logs/input/docker/client_test.go
@@ -5,7 +5,7 @@
 
 // +build docker
 
-package container
+package docker
 
 import (
 	"testing"
