+++ pkg/logs/input/docker/container.go
@@ -5,7 +5,7 @@
 
 // +build docker
 
-package container
+package docker
 
 import (
 	"encoding/json"
