+++ pkg/logs/input/docker/container_test.go
@@ -5,7 +5,7 @@
 
 // +build docker
 
-package container
+package docker
 
 import (
 	"testing"
