+++ pkg/logs/input/docker/scanner.go
@@ -5,7 +5,7 @@
 
 // +build docker
 
-package container
+package docker
 
 import (
 	"context"
