+++ pkg/logs/input/docker/tailer.go
@@ -5,7 +5,7 @@
 
 // +build docker
 
-package container
+package docker
 
 import (
 	"context"
